Move tests to tests/unit folder
