+++ tests/unit/userLogin.test.js
@@ -1,7 +1,7 @@
 const request = require('supertest');
-const { app } = require('../index');
+const { app } = require('../../index');
 const mongoose = require('mongoose');
-const User = require('../models/userModel.js');
+const User = require('../../models/userModel.js');
 const jwt = require('jsonwebtoken');
 
 const userLoginTest = () => describe('POST /auth/login', () => {
